Initialize global variables properly.
This provides a paddr for .data section directing it to flash, and
initializes runtime globals from there in C before loading the Rust code.

This avoids having to deal with accidental undefined behavior from rust
and extensive use of MaybeUninit at the global level for correct
behavior.

A change to speculos is required to support this behavior, as currently
it only maps the .text section when loading an app, but it appears to
work for live devices as the whole FLASH section is copied to the .hex
file during the "cargo ledger" build.